Finish part 2.
Was a bit more difficult to create this using a declerative approach,
but was super fun to crack this with not much code, since most of the
code was  borrowed from part 1's code. Could probably
even refactor the fold at the end to be more descriptive, and use the
same method for both part 1 and part 2.

Signed-off-by: Michael Pacheco <git@michaelpacheco.org>