Day-3:rust: Implement 80% of sample code.
Implemented it as a set of HashSets, where each one checks if the other
has keys. This is more performant than a loop of loops where each char
checks a char exists, this not only uses a faster hash function, but
also makes sure there is only one of each char.

Signed-off-by: Michael Pacheco <git@michaelpacheco.org>