Day-3:rust:part-2: WIP: Partial rework of functions.
Instead of passing owned values, now i am sending references to
everything, this means the logic is more memory efficient since im am
not creating copies of values everywhere, the program is referencing the initial
input (or sample)  dataset, which was initialized at the begining of the main
function, which means it will outlive every function and value that is
computed from it. The program functions still mostly return  multiple iterator chains,
that has not changed.

Signed-off-by: Michael Pacheco <git@michaelpacheco.org>