Day-1:misc:rust: Added declarative soltion
This was an interesting way of finishing the problem, since the input is all in memory (as opposed to what i imagine buf lines handling some of that behind the scenes). This means that i only need to send references to everything, which means that the memory cost is not constant as with the imperative approach, this will now be a O(n) space solution. But it is a *much* faster solution because there are no complicated buffering handled by rust std lib.

* Also applied the more liberal use of the ? operator in my code. which greatly simplified it, the python code in comparison looks incredibly wordy, although not as simple as the elixir code, although in in defense of rust, there is no timing code embedded in it, and its is more duck typed (the input is of course of string) which makes reading the rest of the code harder since the elixir solution to get to its compactness is nesting iterators.

* Also the rust code is pretty fast, i have gotten as far down as ~67µs a the declarative run.

Signed-off-by: Michael Pacheco <git@michaelpacheco.org>